backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1
argon2-cffi==23.1.0
cryptography==44.0.0
PyJWT[crypto]==2.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pyotp==2.9.0
qrcode[pil]==8.0
httpx==0.28.1
itsdangerous==2.2.0
apscheduler==3.10.4
pandas==2.2.3
soccerdata>=1.8.0
numpy>=1.26.0
rich>=13.0
